Extracted helper method to util class
